requirements.txt

pypdf2
nltk
autoawq
sentence-transformers==2.7.0
transformers
torch 
torchvision 
bitsandbytes 
peft
accelerate
chromadb
scikit-learn
python-dotenv
streamlit 